usign float instead of bigdecimal thickness

+++ src/main/java/com/iotiq/application/domain/MaterialParameter.java
@@ -6,15 +6,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @Embeddable
 @AllArgsConstructor
 @NoArgsConstructor
 public class MaterialParameter {
-    private BigDecimal thickness;
+    private Float thickness;
     private Integer flexibility;
     private Integer breathability;
     private Integer moistureWicking;
